requirements.txt
pillow==11.1.0
tk==0.1.0 